extensions/ext_archival_diff/requirements.txt
psycopg2-binary==2.9.10
waybackpy==3.0.6
beautifulsoup4==4.14.3
requests==2.32.5
prometheus_client==0.24.1
